requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.26
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
Werkzeug==3.0.1
